renamed to Xfloat, added test
